T442: minor Validation adaptation
